+++ src/main/java/org/papertrail/listeners/guildlisteners/package-info.java
@@ -1,0 +1,4 @@
+/*
+ * Groups listeners related to guild updates
+ */
+package org.papertrail.listeners.guildlisteners;
