+++ src/main/java/org/papertrail/main/FireRun.java
@@ -7,9 +7,9 @@ import java.sql.SQLException;
 import org.papertrail.cleanup.BotKickListener;
 import org.papertrail.database.DatabaseConnector;
 import org.papertrail.listeners.customlisteners.BotInfoListener;
-import org.papertrail.listeners.customlisteners.ServerBoostListener;
 import org.papertrail.listeners.customlisteners.ServerStatListener;
 import org.papertrail.listeners.customlisteners.SetupListener;
+import org.papertrail.listeners.guildlisteners.ServerBoostListener;
 import org.papertrail.listeners.loglisteners.AuditLogListener;
 import org.papertrail.listeners.loglisteners.LogCommandListener;
 import org.papertrail.listeners.memberlisteners.GuildMemberJoinAndLeaveListener;
